+++ plugins/console/panopticon/services/provider.php
@@ -1,0 +1,52 @@
+<?php
+/*
+ * @package   panopticon
+ * @copyright Copyright (c)2023-2024 Nikolaos Dionysopoulos / Akeeba Ltd
+ * @license   https://www.gnu.org/licenses/agpl-3.0.txt GNU Affero General Public License, version 3 or later
+ */
+
+defined('_JEXEC') || die;
+
+use Joomla\CMS\Extension\PluginInterface;
+use Joomla\CMS\Extension\Service\Provider\MVCFactory;
+use Joomla\CMS\Factory;
+use Joomla\CMS\Plugin\PluginHelper;
+use Joomla\DI\Container;
+use Joomla\DI\ServiceProviderInterface;
+use Joomla\Event\DispatcherInterface;
+use Akeeba\Plugin\Console\Panopticon\Command\CommandFactoryInterface;
+use Akeeba\Plugin\Console\Panopticon\Command\CommandFactoryProvider;
+use Akeeba\Plugin\Console\Panopticon\Extension\Panopticon;
+
+return new class implements ServiceProviderInterface {
+	/**
+	 * Registers the service provider with a DI container.
+	 *
+	 * @param   Container  $container  The DI container.
+	 *
+	 * @return  void
+	 *
+	 * @since   1.0.3
+	 */
+	public function register(Container $container)
+	{
+		$container->registerServiceProvider(new MVCFactory('Akeeba\\Component\\Panopticon'));
+		$container->registerServiceProvider(new CommandFactoryProvider());
+
+		$container->set(
+			PluginInterface::class,
+			function (Container $container) {
+				$config  = (array) PluginHelper::getPlugin('console', 'panopticon');
+				$subject = $container->get(DispatcherInterface::class);
+
+				$config['panopticonCLICommandFactory'] = $container->get(CommandFactoryInterface::class);
+
+				$plugin = new Panopticon($subject, $config);
+
+				$plugin->setApplication(Factory::getApplication());
+
+				return $plugin;
+			}
+		);
+	}
+};