+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Akeeba Panopticon Connector for Joomla 4.x and 5.x version 1.0.3
+
+* ✨ Console plugin for CLI commands
+
 # Akeeba Panopticon Connector for Joomla 4.x and 5.x version 1.0.2
 
 * ✨ Collect server information.